# openApiSchema.yml
openapi: 3.0.0
info:
  title: 股票数据管理API
  description: |
    提供股票数据管理、用户认证和用户数据管理功能的API接口。
    包括股票数据的CRUD操作、用户注册登录、股票选择等功能。
  version: '3.0.0'
  contact:
    name: Stockstock
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 使用JWT令牌进行认证，格式为 "token"
  schemas:
...
    Person:
      type: object
      properties:
        id:
          type: integer
          description: 用户ID
        name:
          type: string
          description: 用户名称
        account:
          type: string
          description: 用户账号
        role:
          type: string
          description: 用户角色(user/admin)
        selectedStocks:
          type: string
          description: 用户选择的股票代码，以逗号分隔
      required:
        - id
        - name
        - account
        - role
    PersonCreate:
      type: object
      properties:
        name:
          type: string
          description: 用户名称
        account:
          type: string
          description: 用户账号
        password:
          type: string
          description: 用户密码
        role（可选，但均为普通用户，未设置可注册为管理员）:
          type: string
          description: 用户角色(user/admin)
      required:
        - name
        - account
        - password
    PersonUpdate:
      type: object
      properties:
        name:
          type: string
          description: 用户名称
        selectedStocks:
          type: string
          description: 用户选择的股票代码，以逗号分隔
      required:
        - name

    # 认证相关模式
    LoginRequest:
      type: object
      properties:
        account:
          type: string
          description: 用户账号
        password:
          type: string
          description: 用户密码
      required:
        - account
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT令牌
        expiresIn:
          type: integer
          description: 过期时间(秒)
      required:
        - token
        - expiresIn
    NameUpdate:
      type: object
      properties:
        name:
          type: string
          description: 新的用户名
      required:
        - name
    PasswordUpdate:
      type: object
      properties:
        password:
          type: string
          description: 新密码
      required:
        - password
        
    # 股票相关模式
    Stock:
      type: object
      properties:
        id:
          type: integer
          description: 记录ID
        stockid:
          type: string
          description: 股票代码
        date:
          type: string
          description: 日期
        TTM:
          type: number
          format: decimal
          description: 市盈率_TTM
        PE:
          type: number
          format: decimal
          description: 市盈率_静
        PB:
          type: number
          format: decimal
          description: 市净率
        PCF:
          type: number
          format: decimal
          description: 市现率
        baiduindex:
          type: integer
          description: 百度指数
        weibo_cnsenti:
          type: number
          format: decimal
          description: 微博情绪_Cnsenti库
        weibo_dictionary:
          type: number
          format: decimal
          description: 微博情绪_纯金融词典
        marketGDP:
          type: number
          format: decimal
          description: 市场GDP
        marketpopulation:
          type: number
          format: decimal
          description: 市场从业人数
        marketaslary:
          type: number
          format: decimal
          description: 市场平均工资
        ratio_TTM:
          type: number
          format: decimal
          nullable: true
          description: r市盈率_TTM
        ratio_PE:
          type: number
          format: decimal
          nullable: true
          description: r市盈率_静
        ratio_PB:
          type: number
          format: decimal
          nullable: true
          description: r市净率
        ratio_PCF:
          type: number
          format: decimal
          nullable: true
          description: r市现率
        ratio_baiduindex:
          type: number
          format: decimal
          nullable: true
          description: r百度指数
        ratio_weibo_cnsenti:
          type: number
          format: decimal
          nullable: true
          description: r微博情绪_Cnsenti库
        ratio_weibo_dictionary:
          type: number
          format: decimal
          nullable: true
          description: r微博情绪_纯金融词典
        ratio_marketGDP:
          type: number
          format: decimal
          nullable: true
          description: r市场GDP
        ratio_marketpopulation:
          type: number
          format: decimal
          nullable: true
          description: r市场从业人数
        ratio_marketaslary:
          type: number
          format: decimal
          nullable: true
          description: r市场平均工资
        total_asset:
          type: number
          format: decimal
          description: 总资产
        quarterly_asset_growth:
          type: number
          format: decimal
          description: 季度资产增长
        cash_flow_perhold_processed:
          type: number
          format: decimal
          description: 每股经处理现金流
        rfr:
          type: number
          format: decimal
          description: 无风险利率
        smooth_asset_growth:
          type: number
          format: decimal
          description: 平滑资产增长
        close_price:
          type: number
          format: decimal
          description: 收盘价格

      required:
        - stockid
        - date
...
    AddStockRequest:
      type: array
      items:
        type: string
        description: 股票代码
      description: 要添加的股票ID数组
      required:
        - stockId

    # 模型预测相关模式
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误消息
        code:
          type: string
          description: 错误代码
    RowsAffectedResponse:
      type: object
      properties:
        rowsAffected:
          type: integer
          description: 受影响的行数
paths:
...
  /persons/register:
    post:
      tags:
        - 用户认证
      summary: 用户注册
      description: 创建新用户账号
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonCreate'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: 新创建的用户ID
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 账户已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /persons/login:
    post:
      tags:
        - 用户认证
      summary: 用户登录
      description: 验证用户凭据并返回JWT令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # 用户管理相关路径
  /persons/all:
    get:
      tags:
        - 用户管理
      summary: 获取所有用户
      description: 管理员获取所有用户信息
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /persons:
    post:
      tags:
        - 用户管理
      summary: 创建用户
      description: 管理员创建新用户
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: 新创建的用户ID
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      tags:
        - 用户管理
      summary: 获取当前用户信息
      description: 获取当前登录用户的详细信息
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - 用户管理
      summary: 更新当前用户信息
      description: 更新当前登录用户的信息
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowsAffectedResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - 用户管理
      summary: 删除当前用户
      description: 删除当前登录用户的账户
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 删除成功
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /persons/name:
    put:
      tags:
        - 用户管理
      summary: 更新用户名
      description: 更新当前登录用户的名称
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowsAffectedResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /persons/password:
    put:
      tags:
        - 用户管理
      summary: 更新密码
      description: 更新当前登录用户的密码
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowsAffectedResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # 用户股票管理相关路径
  /persons/stocks:
    get:
      tags:
        - 用户股票管理
      summary: 获取用户的股票列表
      description: 获取当前用户选择的股票列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  stocks:
                    type: array
                    items:
                      type: string
                    description: 用户选择的股票代码列表
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - 用户股票管理
      summary: 添加股票到用户列表
      description: 向当前用户的选择列表添加股票
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddStockRequest'
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rowsAffected:
                    type: integer
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /persons/stocks/{stockId}:
    delete:
      tags:
        - 用户股票管理
      summary: 从用户列表移除股票
      description: 从当前用户的选择列表中移除特定股票
      security:
        - bearerAuth: []
      parameters:
        - name: stockId
          in: path
          required: true
          schema:
            type: string
          description: 要移除的股票ID
      responses:
        '200':
          description: 移除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rowsAffected:
                    type: integer
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /persons/stocks/{stockId}/check:
    get:
      tags:
        - 用户股票管理
      summary: 检查股票是否在用户列表中
      description: 检查特定股票是否已添加到当前登录用户的选择列表中
      security:
        - bearerAuth: []
      parameters:
        - name: stockId
          in: path
          required: true
          schema:
            type: string
          description: 要检查的股票ID
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  isInList:
                    type: boolean
                    description: 股票是否在用户列表中
              example:
                isInList: true
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # 股票数据管理相关路径
  /stocks:
    get:
      tags:
        - 股票数据
      summary: 获取所有股票信息
      description: 获取数据库中所有股票记录
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - 股票数据
      summary: 创建股票记录
      description: 添加新的股票数据记录(需管理员权限)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  rowsAffected:
                    type: integer
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - 股票数据
      summary: 删除所有股票记录
      description: 删除数据库中所有股票记录(需管理员权限)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 删除成功
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...